Switching back to double precision.
Single precision sometime gives zero adjoint residual too early
due to round of errors. Not suitable for testing.

+++ defs.hpp
@@ -5,7 +5,7 @@
  * Switch between single (float) and double precision by un-/commenting the corresponding lines. 
  */
 
-typedef float MyReal;
-#define MPI_MyReal MPI_FLOAT
-// typedef double MyReal;
-// #define MPI_MyReal MPI_DOUBLE
+// typedef float MyReal;
+// #define MPI_MyReal MPI_FLOAT
+typedef double MyReal;
+#define MPI_MyReal MPI_DOUBLE
